Create 'components' folder and move Timetracker to it.
