+++ src/components/Timetracker/DailyView/EntryDeletionButton.js
