+++ src/components/Timetracker/DailyView/NoEntries.js
